implementoi käyttäjille omat kategoriat ja korjattu bugeja

+++ src/main/java/fi/opiskelijan/budjettilaskuri/web/KategoriaController.java
@@ -1,5 +1,6 @@
 package fi.opiskelijan.budjettilaskuri.web;
 
+import java.security.Principal;
 import java.util.List;
 
 import fi.opiskelijan.budjettilaskuri.domain.Kategoria;
@@ -10,6 +11,8 @@ import fi.opiskelijan.budjettilaskuri.repository.KayttajaRepository;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -28,21 +31,30 @@ public class KategoriaController {
     // GET kategoriat
     @GetMapping("/kategoriat")
     public String listaKategoriat(Model model, Principal principal) {
-        List<Kategoria> kategoriat = kategoriaRepository.findAll();
+        List<Kategoria> kategoriat = kategoriaRepository.findByKayttajaUsername(principal.getName());
         model.addAttribute("kategoriat", kategoriat);
+        model.addAttribute("uusiKategoria", new Kategoria());
+        
         return "kategoriat";
     }
 
-    @PostMapping("/api/kategoriat")
-    public String lisaaKategoria(@RequestBody Kategoria kategoria) {
+    @PostMapping("/kategoriat")
+    public String lisaaKategoria(@ModelAttribute Kategoria kategoria, Principal principal) {
+        Kayttaja user = kayttajaRepository.findByUsername(principal.getName());
+        kategoria.setKayttaja(user);
+
         kategoriaRepository.save(kategoria);
         return "redirect:/kategoriat";
         
     }
 
     @PostMapping("/kategoriat/{id}/poista")
-    public String poistaKategoria(@RequestBody Long id) {
-        kategoriaRepository.deleteById(id);
+    public String poistaKategoria(@PathVariable Long id, Principal principal) {
+        Kategoria kategoria = kategoriaRepository.findById(id).orElse(null);
+        
+        if (kategoria != null && kategoria.getKayttaja().getUsername().equals(principal.getName())) {
+            kategoriaRepository.deleteById(id);
+        }
         return "redirect:/kategoriat";
     }
